Added FASTOpenMDAONanInInputFile as new exception
